pyproject.toml
[tool.poetry]
name = "our-2-brain-cells"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
discord = "^1.0.1"
"discord.py" = "^1.3.4"
pytz = "^2020.1"
youtube-dl = "^2020.7.28"
python-decouple = "^3.3"
googletrans = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
